requirements.txt
firebase-admin
requests
python-dotenv
pandas
pytest
